FIX: In GenomeDescription, chromosomes length now use long type.
git-svn-id: https://subversion.biologie.ens.fr/svn/leburon/trunk/eoulsan@2086 5a62d92b-e016-0410-9103-b89fe532bb34

+++ src/main/java/fr/ens/transcriptome/eoulsan/bio/readsmappers/SimpleGenomeIndexStorage.java
@@ -41,7 +41,7 @@ public class SimpleGenomeIndexStorage implements GenomeIndexStorage {
 
     String genomeName;
     int sequences;
-    int length;
+    long length;
     String genomeMD5;
     String mapperName;
     DataFile file;
@@ -191,7 +191,7 @@ public class SimpleGenomeIndexStorage implements GenomeIndexStorage {
       writer.append("\t");
       writer.append(Integer.toString(ie.sequences));
       writer.append("\t");
-      writer.append(Integer.toString(ie.length));
+      writer.append(Long.toString(ie.length));
       writer.append("\t");
       writer.append(ie.mapperName);
       writer.append("\t");
